Extract env runtime orchestration into internal runtime service
